Add Support for a default HTML.js
This change introduces a barebones HTML file as a fallback for sites
without one.

It also moves the default _template file to match the structure of the
user's site. This way webpack does all the heavy lifting and the default
messages are closer to the default templates.
