+++ lib/isomorphic/create-routes.js
@@ -1,5 +1,4 @@
 import filter from 'lodash/filter'
-import includes from 'lodash/includes'
 import sortBy from 'lodash/sortBy'
 import last from 'lodash/last'
 import { prefixLink } from './gatsby-helpers'
@@ -12,26 +11,9 @@ module.exports = (files, pagesReq) => {
     file.file.name === '_template'
   )
 
-  let rootTemplate
-  try {
-    rootTemplate = pagesReq('./_template')
-  } catch (error) {
-    if (includes(error.message, 'Cannot find module')) {
-      rootTemplate = require('./defaults/_template')
-      // FIXME link to a tutorial/guide/docs/something
-      const defaultTemplateMessage = `
-      Gatsby is currently using the default _template. You can override it by
-      creating a React component at "/pages/_template.js".
-      `
-      console.info(defaultTemplateMessage)
-    } else {
-      throw error
-    }
-  }
-
   const routes = {
     path: prefixLink('/'),
-    component: rootTemplate,
+    component: require('pages/_template'),
     childRoutes: [],
     indexRoute: {},
     pages,
