--- test/fixtures/path-rewrite/html.js
@@ -1,29 +1,0 @@
-import React from 'react'
-import { prefixLink } from 'gatsby-helpers'
-
-module.exports = React.createClass({
-  displayName: 'HTML',
-  propTypes: {
-    body: React.PropTypes.string,
-  },
-  render () {
-    const { body } = this.props
-
-    return (
-      <html lang="en">
-        <head>
-          <meta charSet="utf-8" />
-          <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
-          <meta
-            name="viewport"
-            content="width=device-width, initial-scale=1.0 maximum-scale=5.0"
-          />
-        </head>
-        <body>
-          <div id="react-mount" dangerouslySetInnerHTML={{ __html: body }} />
-          <script src={prefixLink('/bundle.js')} />
-        </body>
-      </html>
-    )
-  },
-})
